I freaking love enums, added cooler banner, updated CANdle template, updated docs [AK]
