+++ src/main/java/frc/lib/BobcatLib/CANdle/BuiltInAnimations.java
@@ -3,6 +3,7 @@ package frc.lib.BobcatLib.CANdle;
 import frc.lib.BobcatLib.Annotations.SeasonBase;
 
 @SeasonBase
+//TODO this is stupid
 public enum BuiltInAnimations {
         ColorFlow,
         Fire,
