+++ src/main/java/frc/lib/BobcatLib/CANdle/CANdle.java
@@ -40,6 +40,7 @@ public class CANdle extends SubsystemBase {
 
   }
 
+  
   /**
    * THIS WILL NOT AUTOMATICALLY TURN OFF, it will persist untill you set it again
    *
