--- src/main/java/frc/lib/BobcatLib/Swerve/README.md
@@ -1,8 +1,0 @@
-Welcome to BobcatLib!
-
-**BobcatEstimator & BobcatSwerveEstimator**
-
-custom versions of WPIlib's PoseEstimator and SwervePoseEstimator, current features include:
-
-* Manual odometry measurment adding
-* dynamic odometry gains
