+++ src/main/java/frc/lib/BobcatLib/Sysid/SysidTest.java
@@ -1,8 +1,3 @@
 package frc.lib.BobcatLib.Sysid;
 
-public enum SysidTest {
-    QUASISTATIC_FORWARD,
-    QUASISTATIC_BACKWARD,
-    DYNAMIC_FORWARD,
-    DYNAMIC_BACKWARD
-}
+
